backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.30.1
httpx==0.27.0
pdfplumber==0.10.3
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3